feat: update launch settings to match prev extension, update build
Talking with Kai and Rob, something we want to do is make the PWA extension
a drop-in replacement for the existing two Chrome and Node debug extensions.
Therefore we want to have the PWA extension be able to read and support the
existing configuration settings, and be invoked by launching the `chrome`
and `node` types. More work, coming soon in a followup PR, will
be needed to actually deal with these settings.

This also moved the build to a Gulp system, supporting NLS and matching
the existing extensions. Part of the build is the ability to have the types
and commands be prefixed with `pwa-`: running `gulp` will generate an extension
that registers types `pwa-node` and `pwa-chrome`, while `gulp --drop-in` will
remove these prefixes. This lets us develop and debug the extension using the
existing packages, until such time as we want to start dogfooding it.