fix - tests hanging if nvm is present
This is definitely the second most annoying bug I've run into in the
last few months. I wanted to wite some integration tests for Node.js.
We have some existing test harnesses, so I just created a new one for
Node targets, easy peasy. But then, they didn't run.

Debugging this was a huge pain, and took most of yesterday afternoon.
The test harness isn't the prettiest code (it's test code, whatever)
and we have some indirection. Eventually, I discovered that for some
reason when we tried to start a Node session, we'd end up connecting
to some `npm-cli.js` script, before the one we wanted to. Where was
this coming from?! I assumed I'd messed something up and tried to figure
out what the expected behavior should be...

So I pulled the repo down today on my Windows machine, and lo and behold
it... just worked? I tried on my Linux VM, and it didn't happen there.
But there's more tools for tracing on Linux, and eventually I found out
it's nvm[1] which I use on Linux and OSX,
but it doesn't work on Windows.

It arises due to code I previously flagged as being 'hacky'[2]. What
happens is that we sent the environment variables for the terminal when
we create it, and these _include the debugger hooks_. Then, the nvm
scripts in the user's .*rc file, which query the CLI for the config.
The CLI was then attaching to our debugging server--but (I eventually
figured out) the test harness wasn't built to have multiple top-level
debuggers attached, so effectively it was 'stealing' our debug session.

This PR fixes the issue by running a child process with proper escaping
and behavior matching the existing launcher in node-debug[3]. There is
currently a separate ChildProcessLauncher that's used only in the debug
server; I'll look at unifying the two in a followup commit.

 1. https://github.com/nvm-sh/nvm
 2. https://github.com/microsoft/vscode-pwa/blob/master/src/ui/terminalProgramLauncher.ts#L36
 3. https://github.com/microsoft/vscode-node-debug/blob/47747454bc6e8c9e48d8091eddbb7ffb54a19bbe/src/node/nodeDebug.ts#L1120