feat: implement console, refactor lifecycles slightly
This implements the `console` option, allowing programs to be opened in the
integration debug console, terminal, or external terminal. I also took the
opportunity to look at lifecycles a little more: I've adjusted the program
launchers (now a subprocess launcher and a terminal launcher) to be stateless
and create `IProgram` instances that carry the session state.

This also involves implementing the callback from the subprocess to the parent
one (containing the process ID and runtime metadata). In the debug2
implementation, we evaluate and return an expression via CDP. In here I've
implemented it as a file callback, where the process writes JSON that the
debug adapter reads. I think this might be a little more simple and reliable--
there's a pretty decent chunk of logic in debug2 about trying and retrying to
evaluate the expression over CDP, and it depends on hitting a breakpoint.
Here we write it out in the bootloader and poll the filesystem for it. It
seems to work well, interested to hear any thoughts.

CDP API generator was previously missing method calls, and only generated
event calls and listeners in the typings. It may make sense to move back to
the official package in the future, but for now I've just patched the
generator.