feat: initial implementation of node attach
This doesn't totally work yet, but I figured this set of changes is big enough.
This implements attaching to an existing/running Node process. It refactors
some of the logic from `NodeLauncher` into an abstract `NodeLauncherBase`
class. We bring in the existing process picker and process tree logic, and
then to attach to a process we run the existing watchdog script (with a few
tweaks) against the target. I'm fairly happy with how this structure ended
up looking.

The code in this PR successfully attaches to a running process, handles
breakpoint state and shows the callstack, but some things are still missing.
I'll add those in a separate PR, or this one if they end up being
one-line changes (I have a feeling there's just a few more .enable()
calls I need to make).

For a separate PR is setting environment variables in the attached process
so that child processes are tracked just like the programs we launch ourselves.