feat: attach to child processes of attached children, support stopOnEntry
A similar mechanism was needed for both scenarios, so they're here. The
NodeTarget now has a lifecycle object, and for node launches we hook into
the `initialized()` point to set a breakpoint on the main script file that
the user asked us to run.

We need to be paused to evaluate code, so the NodeAttacher uses the
initialized() point to request a pause, and when the first pause happens it
retrieves telemetry from the child process, and also installs the environment
variables necessary to instrument children. After doing this, any forked
children are detected automagially--you can try this out by running
`node demos/node/parent.js --inspect` and attaching to it; you can view and
debug in the child process.

The thorny point is around uninstallation when we detach; we'll have a bunch
of orphaned options, and subsequently spawned children will then run into some
issues when trying to connect back to the debug server. Ideally we could have
some piece of code that the inspectee runs automatically when the debugger
detaches, but as far as I can see there's no such mechanism in place. We could
try to do this manually when we disconnect, but not all disconnections
will be graceful.

So next I looked for a way to detect and disable inspection if we can find out
the server went away. And any detection needs to be synchronous so that we can
run it inline in our `--required` script, so we can't try to connect to the
debug server. I ended up just using a lease file. It's not perfect, but it
seems to do the trick.