feat: implement logging (#67)
* feat: implement logging

This is inspired by the previous debug adapter in functionality. It
buffers written logs until we attach sinks, as we can log information
before the DAP or extension context is available. It uses the existing
`trace` option, but is now an object with some additonal information
which allows for filtering; since we now log all DAP and CDP messages,
I figured this could be useful during development to cut down on some
noise.

```ts
export interface ILoggingConfiguration {
  /**
   * Configures whether logs are also returned to the debug console.
   * Defaults to false.
   */
  console: boolean;

  /**
   * Configures the level of logs recorded. Defaults to "Verbose".
   */
  level: string;

  /**
   * Configures where on disk logs are written. If this is null, no logs
   * will be written, otherwise the extension log directory (in VS Code) or
   * OS tmpdir (in VS) will be used.
   */
  logFile: string | null;

  /**
   * Configures what types of logs recorded. For instance, `cdp` will log all
   * CDP protocol messages. If this is empty or not provided, tags will not
   * be filtered.
   */
  tags: ReadonlyArray<string>;
}
```

Logs are written to the file as JSON. I'll modify the existing log
viewer tool to be able to read JSON over its previous line-parsing logic.

* fixup! pr comments