feat: option support in browser attach
This Adds filtering by `url`/`urlFilter` to the browser attacher --
previously it'd always connect to the first tab/target, which is not
always desirable if the user has several tabs open.

We also now support timeouts in the browser attacher and launcher,
set the network cache setting on the target page, and send and log
telemetry about the attached page, like the previous debug adapter did.