fix: breakpoints not applying on first line of block statements
In sourcemaps, if you have a function like we do in the webworker
example containing:

```ts
first.onchange = function() {
  console.log('hello');
```

...then `function() {\n  ` is a single segment in the sourcemap. When
we try to set a breakpoint at the start of the line containing
`console.log`, the previous behavior resolved the position at the start
of the containing segment, so the breakpoint would get set right
before the `function()` expression rather than the console.log
statement. This PR updates the positioning to prefer the segment,
rather than previous segment, when resolving positions.

Sourcemaps also use 0-indexing for columns, which we didn't account
for. I'm not sure if this was causing any issues, but I fixed it here
nonetheless. (Perhaps related to #40.)