feat: track vscode lifecycles correctly
Just something we weren't doing. The existing adapter reads the process
ID when it attaches to a target, and so did we (for Node targets). I
shifted around some logic into a `NodeAttacherBase` so that we can
reuse that query in the extension host, and then from there we reuse
the `TerminalProgram` to track the liveness of the vscode instance
by process ID.

Things I tested that work (from a vscode 'parent' debugging a 'chi;d):
 - Halting debugging from inside the parent instance kills the child
 - Closing the child clears the debugging state in the parent (I think
   it's actually a minor bug in the existing debug adapters where
   this doesn't happen...)
 - Restarting works as expected