feat: implement column breakpoints
Self-explanitory. This was the clearest approach I saw; it looks like we
do a lot of work to map back and forth, but in the majority of cases
we'll only ever enter any of these loops once (e.g. a source without
a sourcemap, or a simple source present in one bundle of compiled code.)
Seems to work well, I'll work on getting some unit tests going tomorrow
morning.