feat: allow disabling pause, falling back to predictor
This works pretty well to resolve the vscode slowness (#99). We still
set a breakpoint on sourcemap load, but clear it as soon as the
breakpoint predictor finishes setting thing up for the root path.

We do already have the `_launchBlockers` promise array, but this only
works if we instruct the debugged process to wait for the debugger on
launch, which we don't seem to do with Chrome and I had trouble
trying to make work in vs code's launch sequence.