feat: allow disabling runtime prediction after prediction finishes
This PR fixes most of the the slow startup time while debugging VS Code.
Like I talked about at various points, the issue is that we have an
instrumentation breakpoint that asks the debugger to pause before we
load any sourcemapped script, so that we can ensure breakpoints get set.
For the large VS Code project, this slows things down a great deal.

If the breakpoint predictor is on and working for this workspace, this
actually isn't necessary--and fortunately, it does work for VS Code.
This PR adds an option that disables the runtime breakpoint once the
breakpoint predictor finishes up. This drops the boot time of the VS
Code debug launch from ~30s to ~6s on my machine, and performance
optimizations to the BreakpointPredictor will drop that further.

Right now this is somewhat of a special case. The breakpoint predictor
_should_ work well on _most_ Node.js apps and some Electron apps. It
might be possible to make the predictor smarter in the future.