feat: improve performance of the breakpoint predictor
This commit is a performance pass on the breakpoint predictor. There
are two main changes.

First, it uses ripgrep instead of our Node.js grepping strategy, when
possible. This benefits us in two ways: it avoids having the breakpoint
prediction logic competing for resources with the interim runtime
breakpoint determination logic; and of course it takes advantage of an
incredibly well-tuned, multicore tool.

Because we don't know what platform the tool will run on, we use a
new library[1] to download the appropriate prebuilt binary into a
directory sent in the launch config. (I'll publish that as a module if
we decide this is the direction we want to go). Eric mentioned for VS
that they could consider shipping the win32 binaries inside the vsix.
Regardless, though, if the ripgrep storage dir isn't set, we'll fall
back to the current Node adapter.

Second, we cache resolved sourcemap information based on the file
modified time, and run the resolution in parallel. This nets us another
solid performance improvement.

Overall, I'm seeing the following numbers for predictor runtime.

```
before changes:    ~6000ms
node.js discovery: ~3000ms (+ caching)
ripgrep discovery:  ~800ms (+ caching, ripgrep)
```

 1. https://github.com/microsoft/vscode-ripgrep-runtime