feat: support webpack namespaces in source map overrides (#97)
This involved adding a 'non-capturing group' syntax for the host map
overrides section. I used the RegEx-style `?:*` sequence, which is not
that pretty, I'm open to ideas. Anyway, this works. The default
source map URIs in webpack are actually formed[1] like:

```
webpack://[namespace]/[resourcePath]
```

It just happens the namespace is usually empty for main targets, but
this is configurable is defaults to `lib` for library targets,
hence #60. This PR adds a non-capturing group for the namespace.

Also makes source map overrides its own class, which saves processing
time (we can pre-parse and generate the regexes) and makes it more
readily testable.

 1. https://github.com/webpack/webpack/blob/91a9f97657b34375cd99e8ba8ef13dd795f1e7c3/lib/SourceMapDevToolPlugin.js#L115