feat: implement differential breakpoints
Fairly straightforward: when we get a breakpoint request, we compare
against the existing breakpoints in the file, and don't touch any that
remain.

I also made a modification to have us wait for the initial breakpoint
`set()` before returning the breakpoint response, so that points are
resolved from the outset. VS Code doesn't move around breakpoints after
the first response comes back, so this allows the column or line to be
adjusted as in the v1 adapters.