fix: ensure bp set completes before updates or pauses (#171)
This is a followup to #169 and the discussion in #162. We now store
a deferred 'set complete' on the breakpoint, which we resolve in a
after returning the set response over DAP. We then guard updates around
this promise and skip sending them if we haven't sent the initial
`setBreakpoints` response yet (the response will contain their full
state with any updates that came in).

Also, if we see that we pause on a breakpoint that we haven't sent
back yet, we'll wait a moment for it to try to be set. I don't _think_
this will ever be hit in VS Code, since it sends all breakpoint
updates individually after the initial configuration, but it's there
regardless and seems to work well.