fixup! set breakpoint on first line to allow source mapping
Following Rob's suggestion of what the v1 adapter does. When we set
a breakpoint in a file, we'll also set a breakpoint on the first line
that we consider to be an instrumentation breakpoint. We treat like
the normal `beforeScriptWithSourceMapExecution` breakpoint, waiting
breakpoints to be set (which in the case of babel/ts-node, includes any
necessary adjustment) before continuing on.

This commit also makes a tweak (over the existing changes in the PR)
to avoid mapping breakpoints by path if we see our source is mapped
in-place. The check for this is rather ugly, but it works.