feat: allow configuring bp predict with outFiles, warn on long predict (#175)
> outFiles 2: electric bugaloo!

With several issues including #170 and #164, it's clear that the
breakpoint predictor will cause problems for some users and there needs
to be a way to configure the grepping behavior. This brings back the
"outFiles" option which, like v1, can be pointed to compiled sources to
direct it what to discover. We still default to the entire workspace,
but have a prompt that appears if we see prediction takes more than 10s:

![](https://memes.peet.io/img/19-12-0805231b-ae8a-4045-a32d-1a8ef2b34874.png)