fix: stopOnEntry not working with omitted program
Fixes #201

As referenced in the comment, this isn't the ideal approach because it
was stolen from us by having a bootloader script. It should work for
most people, though, and I also allowed people to set a custom file path
in "stopOnEntry" in case the heuristics don't work (for instance, if
they start a Node program through a shell script where the path is not
present on the command line).