feat: add code lens for npm scripts
Fixes #196. Adds a code lens to debug an npm script. This adapter
currently doesn't support `nodebug` very well, and there are changes
coming to the terminal in #199, so I didn't implement a "Run" action
yet, but that should be straightforward once those are resolved.

By default the code lens appears atop the "script" section of the
package.json from where it opens a quick-pick, but it can also be
shown above individual scripts or hidden.

This also includes a slight modification to the contrib generator for
configuration to add some type safety.