fix: bps that resolve at entry in hot code by url causing pauses
Previously, if you had code

```
1: function add() {
2:   return 2 + 2;
3: }
```

and set a breakpoint on line 2, but the compiled code was:

```
1: // some
2: // random stuff
3: function add() {
```

Then in our entrypoint pause, we'd be notified that we hit both our
entrypoint breakpoint as well as the user breakpoint. Even though we
then adjust the user breakpoint, we were still seeing that and deciding
to stay paused incorrectly.

This adjusts the logic so that we run the pause check after handling
source maps and adjusting breakpoints.