feat: request async stacks lazily to speed up terminal commands
tl;dr of the investigation from #120 was that the big culprit of slower
commands in the terminal were async stack traces. In this PR, we
introduce logic, enabled by default in the terminal, which does not turn
on async stack tracking until a breakpoint is hit or the debugger
pauses, whichever happens first.

This allows standard build commands like Webpack to run much faster if
the developer is using the debugger terminal like any other terminal and
isn't interested in debugging the current task.

On my machine it reduces debug overhead from 120% to 30%:

```
PS > Measure-Command { webpack }

            Normal Terminal: 12.46 / 12.16 / 13.07
	     Debugger Terminal (master): 22.58 / 23.33 / 22.82 (2.2x)
	     Debugger Terminal (this PR): 16.90 / 16.86 / 16.71 (1.3x)
```