fix: edge cases with vue projects (#242)
Vue is different than what we've seen so far in that, for each
component, it generates multiple 'files' to load CSS, HTML, and so on.
There might be the main file `webpack:///src/App.vue`, a CSS import
in `webpack:///src/App?BADF00D`, etc. The previous behavior of the
source collection was to always overwrite sources by their absolute
path when they got parsed, and as a result breakpoints would often
resolve into the wrong, stubbed file.

In this PR we add a heuristic to prefer the shortest URL if we see
multiple source mapped file mapping to the same file on disk. We can't
just discard any URLs with query strings, since often those are totally
valid, but this seems to make the Vue starter work well.