fix: disconnected event ordering, again (#255)
I was focused on Node before, Chrome is different. In Node, the target
disconnecting is generally 1:1 with the program termination. But in
Chrome, the target (tab) closes before the browser does, and there are
a few more events that get sent out.

Additionally, when tearing down DAP we were previously verbosely
signalling that all sources were removed. This happened during the
disposal process, after everything else was torn down, and they weren't
necessary. Remove them.