feat: handle source roots, pathmapping correctly (#261)
* feat: handle source roots, pathmapping correctly

Previously, we didn't do much with path mappings, and were only
vaguely correct around dealing with source roots. (We were defaulting
to what the source-map library did for us, which it turns out is
sometimes destructive, resolving `../foo.ts` with a sourceRoot of `/`
to `/foo.ts`, as if it was its own filesystem).

This PR imports the bulk of its handling from chrome-debug-core, with
some tweaks to fit. One thing it adds is handling absolute source roots
inside of node modules; it defaults to mapping that relative to the root
of the module. We could expose pathMappings in the Node config as well
to let users configure this.

* fixup! tests