feat: assign workspace folders for attaching to processes
Previously if you attached to a process in a multi-root workspace, we'd
fail because we needed a working directory. This adds functionality
(which only works on *nix, unfortunately) that automatically tries to
assign the workspace folder from the working directory of the process.

Also cleans up handling of process IDs a little bit; these can be made
simpler now that we no longer need to support
the legacy debugger protocol.