refactor: ioc browser finders, add support for edge, tests
This PR pulls the browser finder logic into their own set of classes
that are in Inversify, and more easily testible. This PR adds Edge
discovery on Windows and OS X, with unit tests for each of them.

As I started testing, I swapped to execa[1] for child processes, as it
provides many utilities we can use and is (in simple usage) much more
easily testible than raw child_processes and streams are.

 1. https://github.com/sindresorhus/execa