feat: improve completions for debug console objects (#336)
Previously, we were making sure an expression didn't have side effects,
then sending the whole thing over an evaluating it. But the side effect
check was super aggressive, and we wanted to losen it. Chrome can
actually check expressions for side effects, so it would be cool if it
could do it for us. The property enumeration process has potential
side effects, though, so we need to split it out.

In this change, we now first evaluate the expression, with the
`throwOnSideEffect` flag if we think it might have side effects. If it
returns an object reference, then we invoke the enumeration function on
that. If it doesn't, the expression is probably a primitive type, so
enumerate on the expression directory.

Fixes #251