feat: use companion ext for launching browsers in remote
Adds a new extension to launch browsers, used if we're running in
a remote workspace. Rather than registering the Chrome and Edge
launchers in a remote workspace, we register the new
RemoteBrowserLauncher which executes commands for the companion to run,
and sets up a server to listen for connections.

I still need to add some tests and polish to this code. I think there's
some race condition that running in a remote have exposed, and I also
want to implement a `GzippedPipeTransport` to speed things up over the
wire, since CDP can be chatty for large apps. But it works so far,
and this diff is already pretty big.

Companion source: https://github.com/microsoft/vscode-js-debug-companion