fix: always call terminate() on launchers
Previously we removed the launcher after its session disconnected.
However, the lack of CDP connection does not necessarily mean the target
has terminated--in Blazor, the CDP connection is managed independently
of the browser, but we still want to tear it down when the browser ends.

`terminate()` is now defined to be idempotent, which was the case for
Node.js launchers but not for browsers.