Eric/fix more disconnect hangs (#370)
* fix: dap api generation on interfaces which extend from others

The DAP api generation code wasn't properly expanding interfaces which used the "allOf" param, indicating that they either inherit or are a union of other interfaces.

* fix: respect the stack trace formatting params

* fix: decrement the terminate count before firing the target list changed event

In the _disconnect method, we listen for the onTargetListChanged event as a trigger to continue the disconnect, and also check the terminate count at this point. If the termination count is decremented after the call, this can result in a deadlock where we are waiting for the event after all the targets have terminated.