feat: allow usage of $returnValue
Although the return value is not normally accessible to code, from the
CDP perspective it's just an object reference which can be addressed
like any other reference. So with a little bit of legwork, we can make
it work.

In this PR, we preprocess evaluated code looking for the special
$returnValue identifier. When we get it, we trigger logic to 'hoist' the
return value into the global scope by evaluating `globalThis[var] =
this` in the context of the return object, and then replace $returnValue
in the code. We include a timeout to clear the global reference to avoid
polluting the namespace (which could leak memory).

One interesting thing that arises is that we can technically do this for
any internal object. So, for instance, we could similarly allow
accessing the settled value or error on promises.