fix: lifecycle issues restarting node processes
Fixes #364. There were numerous issues:

- We added cancellation tokens for launches, but restarting used the
same launch configuration as the initial boot, which caused restarts
that happened more than a few seconds later to be cancelled.
- Restarting in attachment specifically could cause the "exit" on the
old program to fire before the new one starts, which the adapter
interprets (when running programs hit 0) as a termination
- Some atomicity issues with the existing restart code

This seems to fix them pretty well...