fix: adjust line breakpoints to the first character in their line
This fixes #407. TypeScript is generally fine, but in Babel sourcemaps
they don't include the leading whitespace as part of the first
statement. This can cause us to place the breakpoint on the start of the
previous line, and in _some_ cases the "variance" calculation will
result in the breakpoint being placed on the next statement, since it
sees the original line had a >0 variance.

In this change, we attempt to scan the source files and automatically
move line breakpoints to the first non-whitespace character in the line
they appear on (if any). This causes the location to be resolved
correctly in all cases.