fix: support spaces in bootloader paths where possible
Most js-debug users are on Node 12, which supports spaces in paths. With
this change, we'll assume we can use spaces unless we verify that the
user is on a version of Node prior to 12. Also, we don't discard the
bootloader if we place it in the tmpdir. The OS will do that for us
at some point in the future.

This fixes https://github.com/microsoft/vscode/issues/96375