fix: handle vue components the best we can
> Observed + tested with Vue CLI 3 and 4

Vue does two weird things:

 1. For each .vue file, they flatten directory structure and only output
    `webpack:///foo.vue`, for instance.
 2. They have generated code in the sourcemaps, which do things like
    create CSS and DOM. Ironically, these files _are_ structured
    correctly (e.g. `webpack:///component/foo.vue`) but we don't want
    to break in them.

This PR introduces some special handling those. Since the directory
structure needs to be restored, that entails we have to do some kind of
re-mapping back to files on disk. I've made the search classes we use
for breakpoint prediction reusable for this: the first time we get a Vue
file that needs remapping, we'll scan the workspace and create the
appropriate mapping.

I have added two regexes that, as precisely as I can, target the build
outputs from the Vue CLI to trigger this remapping when necessary, and
to ignore the nested generated Vue files.

Fixes https://github.com/microsoft/vscode-js-debug/issues/239
Refs https://github.com/vuejs/vue/issues/11023