fix: cdp message perf issue
The current method we use to ensure cdp handler task ordering is
extremely slow (e.g. limited to ~65 messages/sec) and can cause major
issues when debugging certain projects.

We now use the same fix for this as playwright which is much more
performant, and also corrected the check for node version, since this is
only a problem in < 11.0.0 so we don't use the fix on stable 12.18.0
(which doesn't need it)